+++ test/unit/specs/scripts/sdk-generation/fixtures/simplified-3D-secure-schema.js
